Move axis-related tests to new package
