+++ src/test/java/net/imagej/axis/DefaultLinearAxisTest.java
@@ -31,13 +31,11 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertNull;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
@@ -46,7 +44,7 @@ import org.junit.Test;
  * 
  * @author Barry DeZonia
  */
-public class DefaultLinearAxisTest extends AbstractMetaTest {
+public class DefaultLinearAxisTest extends AbstractAxisTest {
 
 	@Test
 	public void testDefaultConstructor() {
