+++ src/test/java/net/imagej/axis/ExponentialRecoveryAxisTest.java
@@ -31,64 +31,64 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
 /**
- * Tests {@link RodbardAxis}.
+ * Tests {@link ExponentialRecoveryAxis}.
  * 
  * @author Barry DeZonia
  */
-public class RodbardAxisTest extends AbstractMetaTest {
+public class ExponentialRecoveryAxisTest extends AbstractAxisTest {
 
 	@Test
 	public void testCtor() {
-		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final ExponentialRecoveryAxis axis =
+			new ExponentialRecoveryAxis(Axes.X, "km", 4, 3, 2, 1);
 
-		assertEquals(Axes.Z, axis.type());
-		assertEquals("lp", axis.unit());
-		assertEquals(1, axis.a(), 0);
-		assertEquals(2, axis.b(), 0);
-		assertEquals(3, axis.c(), 0);
-		assertEquals(4, axis.d(), 0);
+		assertEquals(Axes.X, axis.type());
+		assertEquals("km", axis.unit());
+		assertEquals(4, axis.a(), 0);
+		assertEquals(3, axis.b(), 0);
+		assertEquals(2, axis.c(), 0);
+		assertEquals(1, axis.d(), 0);
 		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
 	@Test
 	public void testOtherStuff() {
-		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final LogLinearAxis axis = new LogLinearAxis();
 
-		axis.setA(2.2);
-		axis.setB(3.3);
-		axis.setC(5.5);
-		axis.setD(7.7);
-		assertEquals(2.2, axis.a(), 0);
-		assertEquals(3.3, axis.b(), 0);
-		assertEquals(5.5, axis.c(), 0);
-		assertEquals(7.7, axis.d(), 0);
+		axis.setA(2);
+		axis.setB(3);
+		axis.setC(5);
+		axis.setD(7);
+		assertEquals(2, axis.a(), 0);
+		assertEquals(3, axis.b(), 0);
+		assertEquals(5, axis.c(), 0);
+		assertEquals(7, axis.d(), 0);
 
 		for (int i = 0; i < 100; i++) {
-			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.0001);
+			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.000001);
 		}
 	}
 
 	@Test
 	public void testCopy() {
-		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
-		final RodbardAxis copy = axis.copy();
+		final ExponentialRecoveryAxis axis =
+			new ExponentialRecoveryAxis(Axes.X, "km", 4, 3, 2, 1);
+		final ExponentialRecoveryAxis copy = axis.copy();
 		assertNotSame(axis, copy);
 		assertEquals(axis, copy);
 		assertEquals(axis.hashCode(), copy.hashCode());
 	}
 
-	private double calValue(final double raw, final RodbardAxis axis) {
-		return axis.a() + (axis.b() - axis.a()) /
-			(1 + Math.pow((raw / axis.c()), axis.d()));
+	private double calValue(final double raw, final ExponentialRecoveryAxis axis)
+	{
+		return axis.a() + axis.b() * (1 - Math.exp(axis.c() + (axis.d() * raw)));
 	}
 }
