+++ src/test/java/net/imagej/axis/InverseRodbardAxisTest.java
@@ -31,73 +31,65 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
-import static org.junit.Assert.assertNull;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
 /**
- * Tests {@link PowerAxis}.
+ * Tests {@link InverseRodbardAxis}.
  * 
  * @author Barry DeZonia
  */
-public class PowerAxisTest extends AbstractMetaTest {
+public class InverseRodbardAxisTest extends AbstractAxisTest {
 
 	@Test
-	public void testCtor1() {
-		final PowerAxis axis = new PowerAxis(2);
-
-		assertUnknown(axis);
-		assertNull(axis.unit());
-		assertEquals(0, axis.a(), 0);
-		assertEquals(1, axis.b(), 0);
-		assertEquals(2, axis.c(), 0);
-		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
-	}
-
-	@Test
-	public void testOtherCtor() {
-		final PowerAxis axis = new PowerAxis(Axes.Z, "lp", 1, 2, 3);
+	public void testCtor() {
+		final InverseRodbardAxis axis =
+			new InverseRodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
 
 		assertEquals(Axes.Z, axis.type());
 		assertEquals("lp", axis.unit());
 		assertEquals(1, axis.a(), 0);
 		assertEquals(2, axis.b(), 0);
 		assertEquals(3, axis.c(), 0);
+		assertEquals(4, axis.d(), 0);
 		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
 	@Test
 	public void testOtherStuff() {
-		final PowerAxis axis = new PowerAxis(3);
+		final InverseRodbardAxis axis =
+			new InverseRodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
 
-		axis.setA(2);
-		axis.setB(3);
-		axis.setC(5);
-		assertEquals(2, axis.a(), 0);
-		assertEquals(3, axis.b(), 0);
-		assertEquals(5, axis.c(), 0);
+		axis.setA(2.2);
+		axis.setB(105.5); // NB - B and C carefully chosen for range 0 to 100
+		axis.setC(-3.3);
+		axis.setD(7.7);
+		assertEquals(2.2, axis.a(), 0);
+		assertEquals(105.5, axis.b(), 0);
+		assertEquals(-3.3, axis.c(), 0);
+		assertEquals(7.7, axis.d(), 0);
 
 		for (int i = 0; i < 100; i++) {
-			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.000001);
+			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.0001);
 		}
 	}
 
 	@Test
 	public void testCopy() {
-		final PowerAxis axis = new PowerAxis(Axes.Z, "lp", 1, 2, 3);
-		final PowerAxis copy = axis.copy();
+		final InverseRodbardAxis axis =
+			new InverseRodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final InverseRodbardAxis copy = axis.copy();
 		assertNotSame(axis, copy);
 		assertEquals(axis, copy);
 		assertEquals(axis.hashCode(), copy.hashCode());
 	}
 
-	private double calValue(final double raw, final PowerAxis axis) {
-		return axis.a() + axis.b() * (Math.pow(raw, axis.c()));
+	private double calValue(final double raw, final InverseRodbardAxis axis) {
+		return axis.a() *
+			Math.pow(((raw - axis.b()) / (axis.c() - raw)), (1 / axis.d()));
 	}
 }
