+++ src/test/java/net/imagej/axis/LogLinearAxisTest.java
@@ -31,31 +31,75 @@
  * #L%
  */
 
-package net.imglib2.meta;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotSame;
+import static org.junit.Assert.assertNull;
 
 import org.junit.Test;
 
 /**
- * Tests {@link DefaultTypedAxis}.
+ * Tests {@link LogLinearAxis}.
  * 
  * @author Barry DeZonia
  */
-public class DefaultTypedAxisTest extends AbstractMetaTest {
-
-	private DefaultTypedAxis axis;
+public class LogLinearAxisTest extends AbstractAxisTest {
 
 	@Test
-	public void test1() {
-		axis = new DefaultTypedAxis();
+	public void testDefaultCtor() {
+		final LogLinearAxis axis = new LogLinearAxis();
+
 		assertUnknown(axis);
+		assertNull(axis.unit());
+		assertEquals(0, axis.a(), 0);
+		assertEquals(1, axis.b(), 0);
+		assertEquals(1, axis.c(), 0);
+		assertEquals(1, axis.d(), 0);
+		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
 	@Test
-	public void test2() {
-		axis = new DefaultTypedAxis(Axes.CHANNEL);
-		assertEquals(Axes.CHANNEL, axis.type());
+	public void testOtherCtor() {
+		final LogLinearAxis axis = new LogLinearAxis(Axes.Z, "lp", 1, 2, 3, 4);
+
+		assertEquals(Axes.Z, axis.type());
+		assertEquals("lp", axis.unit());
+		assertEquals(1, axis.a(), 0);
+		assertEquals(2, axis.b(), 0);
+		assertEquals(3, axis.c(), 0);
+		assertEquals(4, axis.d(), 0);
+		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
+	@Test
+	public void testOtherStuff() {
+		final LogLinearAxis axis = new LogLinearAxis();
+
+		axis.setA(2);
+		axis.setB(3);
+		axis.setC(5);
+		axis.setD(7);
+		assertEquals(2, axis.a(), 0);
+		assertEquals(3, axis.b(), 0);
+		assertEquals(5, axis.c(), 0);
+		assertEquals(7, axis.d(), 0);
+
+		for (int i = 0; i < 100; i++) {
+			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.000001);
+		}
+	}
+
+	@Test
+	public void testCopy() {
+		final LogLinearAxis axis = new LogLinearAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final LogLinearAxis copy = axis.copy();
+		assertNotSame(axis, copy);
+		assertEquals(axis, copy);
+		assertEquals(axis.hashCode(), copy.hashCode());
+	}
+
+	private double calValue(final double raw, final LogLinearAxis axis) {
+		return axis.a() + axis.b() * (Math.log(axis.c() + (axis.d() * raw)));
+	}
 }
