+++ src/test/java/net/imagej/axis/PolynomialAxisTest.java
@@ -31,15 +31,13 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
@@ -48,7 +46,7 @@ import org.junit.Test;
  * 
  * @author Barry DeZonia
  */
-public class PolynomialAxisTest extends AbstractMetaTest {
+public class PolynomialAxisTest extends AbstractAxisTest {
 
 	@Test
 	public void testDefaultConstructor() {
