+++ src/test/java/net/imagej/axis/PowerAxisTest.java
@@ -31,61 +31,55 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertNull;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
 /**
- * Tests {@link LogLinearAxis}.
+ * Tests {@link PowerAxis}.
  * 
  * @author Barry DeZonia
  */
-public class LogLinearAxisTest extends AbstractMetaTest {
+public class PowerAxisTest extends AbstractAxisTest {
 
 	@Test
-	public void testDefaultCtor() {
-		final LogLinearAxis axis = new LogLinearAxis();
+	public void testCtor1() {
+		final PowerAxis axis = new PowerAxis(2);
 
 		assertUnknown(axis);
 		assertNull(axis.unit());
 		assertEquals(0, axis.a(), 0);
 		assertEquals(1, axis.b(), 0);
-		assertEquals(1, axis.c(), 0);
-		assertEquals(1, axis.d(), 0);
+		assertEquals(2, axis.c(), 0);
 		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
 	@Test
 	public void testOtherCtor() {
-		final LogLinearAxis axis = new LogLinearAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final PowerAxis axis = new PowerAxis(Axes.Z, "lp", 1, 2, 3);
 
 		assertEquals(Axes.Z, axis.type());
 		assertEquals("lp", axis.unit());
 		assertEquals(1, axis.a(), 0);
 		assertEquals(2, axis.b(), 0);
 		assertEquals(3, axis.c(), 0);
-		assertEquals(4, axis.d(), 0);
 		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
 	}
 
 	@Test
 	public void testOtherStuff() {
-		final LogLinearAxis axis = new LogLinearAxis();
+		final PowerAxis axis = new PowerAxis(3);
 
 		axis.setA(2);
 		axis.setB(3);
 		axis.setC(5);
-		axis.setD(7);
 		assertEquals(2, axis.a(), 0);
 		assertEquals(3, axis.b(), 0);
 		assertEquals(5, axis.c(), 0);
-		assertEquals(7, axis.d(), 0);
 
 		for (int i = 0; i < 100; i++) {
 			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.000001);
@@ -94,14 +88,14 @@ public class LogLinearAxisTest extends AbstractMetaTest {
 
 	@Test
 	public void testCopy() {
-		final LogLinearAxis axis = new LogLinearAxis(Axes.Z, "lp", 1, 2, 3, 4);
-		final LogLinearAxis copy = axis.copy();
+		final PowerAxis axis = new PowerAxis(Axes.Z, "lp", 1, 2, 3);
+		final PowerAxis copy = axis.copy();
 		assertNotSame(axis, copy);
 		assertEquals(axis, copy);
 		assertEquals(axis.hashCode(), copy.hashCode());
 	}
 
-	private double calValue(final double raw, final LogLinearAxis axis) {
-		return axis.a() + axis.b() * (Math.log(axis.c() + (axis.d() * raw)));
+	private double calValue(final double raw, final PowerAxis axis) {
+		return axis.a() + axis.b() * (Math.pow(raw, axis.c()));
 	}
 }
