+++ src/test/java/net/imagej/axis/RodbardAxisTest.java
@@ -31,45 +31,62 @@
  * #L%
  */
 
-package net.imglib2.meta.axis;
+package net.imagej.axis;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
-import static org.junit.Assert.assertNull;
-import net.imglib2.meta.AbstractMetaTest;
-import net.imglib2.meta.Axes;
 
 import org.junit.Test;
 
 /**
- * Tests {@link IdentityAxis}.
+ * Tests {@link RodbardAxis}.
  * 
  * @author Barry DeZonia
  */
-public class IdentityAxisTest extends AbstractMetaTest {
+public class RodbardAxisTest extends AbstractAxisTest {
 
 	@Test
-	public void testVarious() {
-		IdentityAxis axis = new IdentityAxis();
-		assertUnknown(axis);
-		axis = new IdentityAxis(Axes.Y);
-		assertEquals(Axes.Y, axis.type());
-		assertNull(axis.unit());
-		axis.setUnit("BAMBALOOIE");
-		assertEquals("BAMBALOOIE", axis.unit());
-		axis.setType(Axes.CHANNEL);
-		assertEquals(Axes.CHANNEL, axis.type());
-		assertEquals(5, axis.calibratedValue(5), 0);
-		assertEquals(5, axis.rawValue(5), 0);
+	public void testCtor() {
+		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+
+		assertEquals(Axes.Z, axis.type());
+		assertEquals("lp", axis.unit());
+		assertEquals(1, axis.a(), 0);
+		assertEquals(2, axis.b(), 0);
+		assertEquals(3, axis.c(), 0);
+		assertEquals(4, axis.d(), 0);
+		assertEquals(calValue(4, axis), axis.calibratedValue(4), 0);
+	}
+
+	@Test
+	public void testOtherStuff() {
+		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+
+		axis.setA(2.2);
+		axis.setB(3.3);
+		axis.setC(5.5);
+		axis.setD(7.7);
+		assertEquals(2.2, axis.a(), 0);
+		assertEquals(3.3, axis.b(), 0);
+		assertEquals(5.5, axis.c(), 0);
+		assertEquals(7.7, axis.d(), 0);
+
+		for (int i = 0; i < 100; i++) {
+			assertEquals(axis.rawValue(axis.calibratedValue(i)), i, 0.0001);
+		}
 	}
 
 	@Test
 	public void testCopy() {
-		final IdentityAxis axis = new IdentityAxis(Axes.Z);
-		final IdentityAxis copy = axis.copy();
+		final RodbardAxis axis = new RodbardAxis(Axes.Z, "lp", 1, 2, 3, 4);
+		final RodbardAxis copy = axis.copy();
 		assertNotSame(axis, copy);
 		assertEquals(axis, copy);
 		assertEquals(axis.hashCode(), copy.hashCode());
 	}
 
+	private double calValue(final double raw, final RodbardAxis axis) {
+		return axis.a() + (axis.b() - axis.a()) /
+			(1 + Math.pow((raw / axis.c()), axis.d()));
+	}
 }
